requirement.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10